requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
mutagen==1.47.0
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
APScheduler==3.10.4